requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
importlib-metadata==0.18
more-itertools==6.0.0
multidict==4.5.2
olefile==0.46
packaging==19.0
Pillow==5.4.1
pluggy==0.12.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.5
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.7.0
pytest==5.1.0
python-decouple==3.1
requests==2.22.0
rsa==4.0
six==1.12.0
Telethon==1.9.0
urllib3==1.24.3
wcwidth==0.1.7
yarl==1.3.0
zipp==0.5.1
